+++ flask_dance/__init__.py
@@ -1,3 +1,3 @@
 from .consumer import OAuth1ConsumerBlueprint, OAuth2ConsumerBlueprint
 
-__version__ = "3.3.1"
+__version__ = "4.0.0"