+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 * Dropped support for Python 2 and Python 3.5
 * If you are using the SQLAlchemy token storage, this project now depends on
   SQLAlchemy version 1.3.11 and above. `sqlalchemy-utils` is no longer necessary.
+* Added Twitch pre-set configuration
 
 `3.3.1`_ (2021-03-01)
 ---------------------